requirements.txt
pillow>=9.0.0,<10.0.0
numpy>=1.20.0,<1.25.0
opencv-python>=4.5.0
open3d>=0.15.0
torch>=2.0.0,<2.1.0
torchvision>=0.15.0,<0.16.0
